% Generated by build script
\name{swpc_mag}
\alias{swpc_mag}
\title{Fetch NOAA SWPC solar wind magnetic field data}
\usage{
swpc_mag()
}
\arguments{
\item{days}{Time window: 0.08 (2hr), 0.25 (6hr), 1, 3, or 7 (default 7)}
}
\value{tibble: time_tag (POSIXct), bx_gsm, by_gsm, bz_gsm, lon_gsm,}
\description{Fetch NOAA SWPC solar wind magnetic field data}
