% Generated by build script
\name{tides_predictions}
\alias{tides_predictions}
\title{Get tide predictions for a station}
\usage{
tides_predictions()
}
\arguments{
\item{station}{Station ID (e.g. '9414290' for San Francisco)}
\item{begin_date}{Start date YYYYMMDD}
\item{end_date}{End date YYYYMMDD}
}
\value{tibble of tide predictions}
\description{Get tide predictions for a station}
