% Generated by build script
\name{tides_water_levels}
\alias{tides_water_levels}
\title{Get water levels for a station}
\usage{
tides_water_levels()
}
\arguments{
\item{station}{Station ID}
\item{begin_date}{Start date YYYYMMDD}
\item{end_date}{End date YYYYMMDD}
}
\value{tibble of water levels}
\description{Get water levels for a station}
